+++ src/constants/communityEmoji.ts
@@ -1,13 +1,13 @@
 import RedHeartEmojiImg from '@/public/images/RedHeartEmoji.png';
 import HappyEmojiImg from '@/public/images/HappyEmoji.png';
-import ExitedEmojiImg from '@/public/images/ExitedEmoji.png';
+import ExcitedEmojiImg from '@/public/images/ExcitedEmoji.png';
 import AngryEmojiImg from '@/public/images/AngryEmoji.png';
 import CryEmojiImg from '@/public/images/CryEmoji.png';
 
 export const EMOJI_ICON = {
   heart: RedHeartEmojiImg,
   smile: HappyEmojiImg,
-  exited: ExitedEmojiImg,
+  excited: ExcitedEmojiImg,
   angry: AngryEmojiImg,
   crying: CryEmojiImg,
 };
